Create Get-AzureLogs
Not exactly the cleanest code and it is prone to some possible errors.  This assumes you have manually pulled down your latest set of logs from Azure.  It will then only pull the logs from the last 15 minutes and remove the oldest ones based on your specification (currently 14 days or older).  It writes to a log file for the individual adds and removals and includes a tally of the removals.  If you wanted to tally the adds, just include a similar variable in the NewBlobs foreach loop and perform a similar LogWrite as on Line 53.  If you have a ton of logs to pull, this script will be a cpu hog.  You may want to do something like Cloudberry to bring in your first pull.  The Get-AzureStorageBlobContent has -ConcurrentTaskCount<Int32> which can assist in throttling the CPU usage for the task. 

So how do you get the Logs into Azure Blob Storage to begin with?  Tom Hollander's blog on msdn helped a lot http://blogs.msdn.com/b/tomholl/archive/2013/05/09/zip-your-iis-log-files-before-transferring-with-windows-azure-diagnostics.aspx