Powershell Empty Folder cleanup
Used this after cleaning up a music library.  It does not appear clean
up all the empty folders, it could be due to hidden files preventing the
deletion, but it does throw an error.   It does clean up a good majority
of them.

Add -whatif to line 23 if you want to see what it will remove.  You can
also take the $emptydirs value and pipe it to an export-csv to see the
files you will be removing.