Update and rename Get-AzureLogs to Get-AzureLogs.ps1
Added conditional statements for logging if blobs are deleted or not.  Also set the LogTime to just a date rather than the time.  This will then create a daily log file and append to that throughout the day.  It will cutdown heavily on small log files.  As the task runs every 15 minutes, these files will add up.  If you run it more frequently then you see the problem.