Update Get-AzureLogs.ps1
Added mandatory parameters to use when running the script.  This allows a bit more flexibility if you would like to reuse on multiple environments.  Simply run it as a scheduled task with the appropriate arguments.  This also allows you to remove the Azure Storage keys from the script in case you want to share it and you don't have to worry as much about sanitizing it.  I may also consider adding arguments for the rest of the Context specific items.

Also make sure your log service account has access to the dump location <facepalm>