Handling file overwrites
A problem I ran into with a scheduled task is that if a log was already pulled from the current day, then when the task ran again that day, the script would require confirmation on overwriting a file.  Adding "-Force" on line 52 will allow the script to overwrite any existing files.  In this case it is ok as the files will either be the same, or contain more data depending on how IIS is logging.  In most cases they will be the same since the zipped logs will be created when the task runs on the Azure instances.